+++ crates/icp-cli/src/commands/args.rs
@@ -13,7 +13,7 @@ use crate::options::{EnvironmentOpt, IdentityOpt, NetworkOpt};
 #[derive(Args, Debug)]
 pub(crate) struct CanisterCommandArgs {
     // Note: Could have flattened CanisterEnvironmentArg to avoid adding child field
-    /// Name or principal of canister to target
+    /// Name or principal of canister to target.
     /// When using a name an environment must be specified.
     pub(crate) canister: Canister,